Fixed more errors and the begining of Test.

+++ Studynomic3/app/src/main/java/com/opazoweb/studynomic3/data/calculators/WorkCalculator.kt
@@ -51,7 +51,7 @@ class WorkCalculator(
     }
 
     private fun leaveToDeci(leave: Double):Double {
-        return (100 - leave.toDouble()) / 100
+        return (100 - leave) / 100
     }
 
 }
